+++ core/src/main/java/edu/pnu/movement/RandomWayPoint.java
@@ -61,7 +61,7 @@ public class RandomWayPoint extends AbstractWayPoint {
         return coordinates.get(randNumber);
     }
     
-    public Coordinate getNext(MovingObject mo, long time) {
+    public Coordinate getNext(MovingObject mo, double time) {
         if(finder == null) {
             finder = new DijkstraPathFinder(graph);
